+++ pkgs/ffigen/lib/src/code_generator/writer.dart
@@ -198,8 +198,9 @@ const _\$objcVersionCheck = $objcPrefix.ObjCVersionCheck(
         'chosen by the most common compilers for the various OS and '
         'architecture combinations. To prevent any crashes, remove the '
         'enums from your API surface. To rely on the (unsafe!) mimicking, '
-        'you can silence this warning by adding silence-enum-warning: true '
-        'to the FFIgen config. Affected enums:\n\t${names.join('\n\t')}',
+        'you can silence this warning by setting node.silenceWarning = true '
+        'in a Visitor(enumClass: ...). Affected enums:'
+        '\n\t${names.join('\n\t')}',
       );
     }