+++ pkgs/ffigen/test/header_parser_tests/dart_handle_test.dart
@@ -22,7 +22,7 @@ void main() {
 ${strings.name}: 'NativeLibrary'
 ${strings.description}: 'Dart_Handle Test'
 ${strings.output}: 'unused'
-${strings.compilerOpts}: '-I${path.join(sdkPath, "include")}'
+${strings.compilerOpts}: '-I${path.join(sdkPath!, "include")}'
 
 ${strings.headers}:
   ${strings.entryPoints}:
@@ -33,9 +33,9 @@ ${strings.headers}:
         ),
       );
     });
-    test('Expected Bindings', () {
+    test('Expected Bindings', () async {
       final context = testContext();
-      matchLibraryWithExpected(
+      await matchLibraryWithExpected(
         context,
         actual,
         'header_parser_dart_handle_test_output.dart',