+++ pkgs/ffigen/test/header_parser_tests/imported_types_test.dart
@@ -30,9 +30,9 @@ ${strings.headers}:
         ),
       );
     });
-    test('Expected Bindings', () {
+    test('Expected Bindings', () async {
       final context = testContext();
-      matchLibraryWithExpected(
+      await matchLibraryWithExpected(
         context,
         actual,
         'header_parser_imported_types_test_output.dart',